requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-sqlite
pandas==2.1.1